pyproject.toml
[tool.poetry]
name = "walletapi"
version = "0.1.0"
description = "p-project"
authors = ["maniakalochka <prochorwolochoff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
pydantic = {extras = ["email"], version = "^2.9.2"}
pytest = "^8.3.3"
alembic = "^1.14.0"
pytest-asyncio = "^0.24.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.36"}
slugify = "^0.0.1"
uvicorn = "^0.32.0"
black = "^24.10.0"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
python-multipart = "^0.0.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.2"
flake8 = "^7.1.1"
mypy = "^1.13.0"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]
